# Simplify some `Iterator` methods.