@nnethercote 2019-09-18
@simulacrum: time to test the new all-in-one perf CI command:

@bors try @rust-timer queue