@rkruppe 2019-09-18
Note that using internal iteration over `for` loops in all of these iterator "sinks" was done not out of cleverness but because internal iteration tends to optimize much better for some iterators (e.g., anything involving `chain`). See for example the benchmark cited in https://github.com/rust-lang/rust/pull/45595 for `sum`. It's very bad that this turns out to be worse for the iterations that are common in rustc and I won't try to block this PR, but I think we should expect that this causes regressions for other kinds of code. Ideally we could have our cake and eat it too.