@nnethercote 2019-09-18
@rkruppe: I read through #45595 but I don't see an explanation of the `sum` benchmark, and intuitively I'm struggling to understand how it could optimize better. Can you explain some more?

I'd be interested in concrete examples of where this change causes slowdowns, because the perf results for rustc itself are *ridiculously good*. A 48.8% instruction count win? A 50.5% wall-time win? That's crazy. I had only tested with `Check` runs of `keccak` and `inflate`, because they are benchmarks where I know that the slowness of `all`, and they gave the slight speed-ups I expected. I will double check this with local runs and try to work out why it's such a big improvement.