@Mark-Simulacrum 2019-09-18
Hm, so I wonder if https://github.com/rust-lang/rust/pull/62429 is somewhat related here. I'd be interested in seeing if we could get similar results with the old code but replacing e.g. `|_| true` with `true_bool` with `fn true_bool() -> bool { true }` or so. But I do think it makes sense to just land this if there are no objections given the wins here :)