@nnethercote 2019-09-19
Here is a test program that demonstrates @rkruppe's point:
```rust
// bench.rs
// - Compile with `rustc --test -O bench.rs`
// - Run with `./bench --bench`
#![feature(test)]

extern crate test;

use test::Bencher;

#[bench]
fn long_all(b: &mut Bencher) {
    b.iter(|| {
        let n = test::black_box(100000);
        let k = test::black_box(1000000);
        (0..2*n).all(|x| x < k)
    })
}

#[bench]
fn long_all_chained(b: &mut Bencher) {
    b.iter(|| {
        let n = test::black_box(100000);
        let k = test::black_box(1000000);
        (0..n).chain(0..n).all(|x| x < k)
    })
}
```
Before:
```
running 2 tests
test long_all         ... bench:      47,849 ns/iter (+/- 2,881)
test long_all_chained ... bench:      49,985 ns/iter (+/- 4,589)
```
After:
```
running 2 tests
test long_all         ... bench:      47,247 ns/iter (+/- 3,094)
test long_all_chained ... bench:     142,170 ns/iter (+/- 5,201)
```
So we have a demonstrated slowdown for a `chain`ed use in a microbenchmark. It's also interesting that this microbenchmark doesn't show a speed-up for `long_all`, which is counter to the godbolt microbenchmark, and counter to the real-world experience in rustc itself.